A profile for controlling the gui
The default set is tuned for moving windows around with the
keyboard and navigating with the mouse. All of the keyboard
modifiers CTRL, ALT, SHIFT, SUPER are present. Along with the
arrow keys, there is a lot of utility there by default in most GUI's.

There is also access to all of the characters on a US layout QWERTY
keyboard for those bold enough to learn text entry.